Rename script, add check for migrations and use in travis
